Move some classes to common
* Should be used by everything
* Core should only be used by core elements
