+++ source/common/lazy.h
