+++ source/common/nocopyormove.h
