+++ tests/test_nocopyormove.cpp
@@ -1,12 +1,16 @@
+#define CATCH_CONFIG_RUNTIME_STATIC_REQUIRE
 #include <catch2/catch_all.hpp>
+#include <type_traits>
 
-#include "core/nocopyormove.h"
+#include "common/nocopyormove.h"
+
+CATCH_STATIC_REQUIRE_FALSE(std::is_move_constructible<NoCopyOrMove>::value);
 
 TEST_CASE("no copy", "[Util]") {
     NoCopyOrMove nc;
 
     /*SECTION("can be moved constructed") {
-        NoCopyOrMove nc1(std::move(nc));
+        NoCopyOrMove nc1(std::move(NoCopyOrMove());
     }
 
     SECTION("can be moved") {
